+++ api/database_cluster.go
@@ -4,25 +4,25 @@ import "github.com/labstack/echo/v4"
 
 // CreateDatabaseCluster creates a new db cluster inside the given k8s cluster.
 func (e *EverestServer) CreateDatabaseCluster(ctx echo.Context, kubernetesID string) error {
-	return e.proxyKubernetes(ctx, kubernetesID, "")
+	return e.everestK8s.ProxyKubernetes(ctx, kubernetesID, "")
 }
 
 // ListDatabaseClusters List of the created database clusters on the specified kubernetes cluster.
 func (e *EverestServer) ListDatabaseClusters(ctx echo.Context, kubernetesID string) error {
-	return e.proxyKubernetes(ctx, kubernetesID, "")
+	return e.everestK8s.ProxyKubernetes(ctx, kubernetesID, "")
 }
 
 // DeleteDatabaseCluster Create a database cluster on the specified kubernetes cluster.
 func (e *EverestServer) DeleteDatabaseCluster(ctx echo.Context, kubernetesID string, name string) error {
-	return e.proxyKubernetes(ctx, kubernetesID, name)
+	return e.everestK8s.ProxyKubernetes(ctx, kubernetesID, name)
 }
 
 // GetDatabaseCluster Get the specified database cluster on the specified kubernetes cluster.
 func (e *EverestServer) GetDatabaseCluster(ctx echo.Context, kubernetesID string, name string) error {
-	return e.proxyKubernetes(ctx, kubernetesID, name)
+	return e.everestK8s.ProxyKubernetes(ctx, kubernetesID, name)
 }
 
 // UpdateDatabaseCluster Replace the specified database cluster on the specified kubernetes cluster.
 func (e *EverestServer) UpdateDatabaseCluster(ctx echo.Context, kubernetesID string, name string) error {
-	return e.proxyKubernetes(ctx, kubernetesID, name)
+	return e.everestK8s.ProxyKubernetes(ctx, kubernetesID, name)
 }
