+++ api/everest.go
@@ -4,28 +4,10 @@ package api
 //go:generate ../bin/oapi-codegen --config=server.cfg.yml  ../docs/spec/openapi.yml
 
 import (
-	"context"
-	"fmt"
-	"net/http"
-	"net/http/httputil"
-	"net/url"
-	"strings"
-
-	"github.com/AlekSi/pointer"
-	"github.com/labstack/echo/v4"
-	everestv1alpha1 "github.com/percona/everest-operator/api/v1alpha1"
-	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
 	"go.uber.org/zap"
-	corev1 "k8s.io/api/core/v1"
-	k8serr "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
 
 	"github.com/percona/percona-everest-backend/cmd/config"
 	"github.com/percona/percona-everest-backend/model"
-	perconak8s "github.com/percona/percona-everest-backend/pkg/kubernetes"
 )
 
 const (
@@ -39,6 +21,7 @@ type EverestServer struct {
 	l              *zap.SugaredLogger
 	storage        storage
 	secretsStorage secretsStorage
+	everestK8s     everestK8s
 }
 
 // List represents a general object with the list of items.
@@ -52,12 +35,12 @@ func NewEverestServer(c *config.EverestConfig, l *zap.SugaredLogger) (*EverestSe
 		config: c,
 		l:      l,
 	}
-	err := e.initStorages()
+	err := e.initEverest()
 
 	return e, err
 }
 
-func (e *EverestServer) initStorages() error {
+func (e *EverestServer) initEverest() error {
 	db, err := model.NewDatabase(pgStorageName, e.config.DSN, pgMigrationsDir)
 	if err != nil {
 		return err
@@ -65,190 +48,6 @@ func (e *EverestServer) initStorages() error {
 	e.storage = db
 	e.secretsStorage = db // so far the db implements both interfaces - the regular storage and the secrets storage
 	_, err = db.Migrate()
+	e.everestK8s = newEverestK8s(e.storage, e.secretsStorage, e.l)
 	return err
-}
-
-// applyObjectStorage creates k8s objects in the given k8s cluster.
-func (e *EverestServer) applyObjectStorage(ctx echo.Context, bs BackupStorage, secretFields map[string]string, kubernetesID string) error {
-	k, err := e.storage.GetKubernetesCluster(ctx.Request().Context(), kubernetesID)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{Message: pointer.ToString(err.Error())})
-	}
-
-	everestClient, err := perconak8s.NewFromSecretsStorage(
-		ctx.Request().Context(), e.secretsStorage, k.ID,
-		k.Namespace, logrus.NewEntry(logrus.StandardLogger()),
-	)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{Message: pointer.ToString(err.Error())})
-	}
-
-	secretName := buildSecretName(bs.Name)
-	secret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      secretName,
-			Namespace: k.Namespace,
-		},
-		StringData: secretFields,
-		Type:       corev1.SecretTypeOpaque,
-	}
-	_, err = everestClient.CreateSecret(ctx.Request().Context(), secret)
-	// if such Secret is already present in k8s - consider it as created and do nothing (fixme)
-	if err != nil && !k8serr.IsAlreadyExists(err) {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{Message: pointer.ToString(err.Error())})
-	}
-
-	var url string
-	if bs.Url != nil {
-		url = *bs.Url
-	}
-	backupStorage := &everestv1alpha1.ObjectStorage{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      bs.Name,
-			Namespace: k.Namespace,
-		},
-		Spec: everestv1alpha1.ObjectStorageSpec{
-			Type:                  everestv1alpha1.ObjectStorageType(bs.Type),
-			Bucket:                bs.BucketName,
-			Region:                bs.Region,
-			EndpointURL:           url,
-			CredentialsSecretName: secretName,
-		},
-	}
-
-	err = everestClient.CreateObjectStorage(ctx.Request().Context(), backupStorage)
-	// if such ObjectStorage is already present in k8s - consider it as created and do nothing (fixme)
-	if err != nil && !k8serr.IsAlreadyExists(err) {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{Message: pointer.ToString(err.Error())})
-	}
-
-	return nil
-}
-
-// removeObjectStorage removes k8s objects from the given k8s cluster.
-func (e *EverestServer) removeObjectStorage(ctx echo.Context, storageName, kubernetesID string) error {
-	k, err := e.storage.GetKubernetesCluster(ctx.Request().Context(), kubernetesID)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{Message: pointer.ToString(err.Error())})
-	}
-
-	everestClient, err := perconak8s.NewFromSecretsStorage(
-		ctx.Request().Context(), e.secretsStorage, k.ID,
-		k.Namespace, logrus.NewEntry(logrus.StandardLogger()),
-	)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{Message: pointer.ToString(err.Error())})
-	}
-
-	dbClusters, err := e.getDBClustersByObjectStorage(ctx.Request().Context(), everestClient, storageName)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{Message: pointer.ToString(err.Error())})
-	}
-
-	if err = buildObjectStorageInUseError(dbClusters, storageName); err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusBadRequest, Error{Message: pointer.ToString(err.Error())})
-	}
-
-	err = everestClient.DeleteObjectStorage(ctx.Request().Context(), storageName, k.Namespace)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{Message: pointer.ToString(err.Error())})
-	}
-
-	secretName := buildSecretName(storageName)
-	err = everestClient.DeleteSecret(ctx.Request().Context(), secretName, k.Namespace)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{Message: pointer.ToString(err.Error())})
-	}
-
-	return nil
-}
-
-func (e *EverestServer) proxyKubernetes(ctx echo.Context, kubernetesID, resourceName string) error {
-	cluster, err := e.storage.GetKubernetesCluster(ctx.Request().Context(), kubernetesID)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{Message: pointer.ToString(err.Error())})
-	}
-	encodedSecret, err := e.secretsStorage.GetSecret(ctx.Request().Context(), kubernetesID)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusInternalServerError, Error{Message: pointer.ToString(err.Error())})
-	}
-
-	config, err := clientcmd.BuildConfigFromKubeconfigGetter("", newConfigGetter(encodedSecret).loadFromString)
-	if err != nil {
-		e.l.Error(err)
-		return ctx.JSON(http.StatusBadRequest, Error{Message: pointer.ToString(err.Error())})
-	}
-	reverseProxy := httputil.NewSingleHostReverseProxy(
-		&url.URL{
-			Host:   strings.TrimPrefix(config.Host, "https://"),
-			Scheme: "https",
-		})
-	transport, err := rest.TransportFor(config)
-	if err != nil {
-		return ctx.JSON(http.StatusBadRequest, Error{Message: pointer.ToString(err.Error())})
-	}
-	reverseProxy.Transport = transport
-	req := ctx.Request()
-	req.URL.Path = buildProxiedURL(ctx.Request().URL.Path, kubernetesID, resourceName, cluster.Namespace)
-	reverseProxy.ServeHTTP(ctx.Response(), req)
-	return nil
-}
-
-func (e *EverestServer) getDBClustersByObjectStorage(ctx context.Context, everestClient *perconak8s.Kubernetes, storageName string) ([]everestv1alpha1.DatabaseCluster, error) {
-	list, err := everestClient.ListDatabaseClusters(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	dbClusters := make([]everestv1alpha1.DatabaseCluster, 0, len(list.Items))
-	for _, dbCluster := range list.Items {
-		for _, schedule := range dbCluster.Spec.Backup.Schedules {
-			if schedule.ObjectStorageName == storageName {
-				dbClusters = append(dbClusters, dbCluster)
-				break
-			}
-		}
-	}
-
-	return dbClusters, nil
-}
-
-func buildObjectStorageInUseError(dbClusters []everestv1alpha1.DatabaseCluster, storageName string) error {
-	if len(dbClusters) == 0 {
-		return nil
-	}
-	names := make([]string, 0, len(dbClusters))
-	for _, cluster := range dbClusters {
-		names = append(names, cluster.Name)
-	}
-
-	return errors.Errorf("the ObjectStorage '%s' is used in following DatabaseClusters: %s. Please update the DatabaseClusters configuration first", storageName, strings.Join(names, ","))
-}
-
-func buildProxiedURL(uri, kubernetesID, resourceName, namespace string) string {
-	// cut the /kubernetes part
-	uri = strings.TrimPrefix(uri, "/v1/kubernetes/"+kubernetesID)
-
-	// cut the resource name if present
-	uri = strings.TrimSuffix(uri, resourceName)
-
-	// remove kebab-case
-	uri = strings.ReplaceAll(uri, "-", "")
-	return fmt.Sprintf("/apis/everest.percona.com/v1alpha1/namespaces/%s%s%s", namespace, uri, resourceName)
-}
-
-func buildSecretName(storageName string) string {
-	return storageName + "-secret"
 }
